requirements.txt
appdirs==1.4.4
openai==0.28.0
pre-commit==3.4.0
pydantic==2.3.0
python-dotenv==1.0.0
zstandard==0.21.0
